+++ insanic/tracing/utils.py
@@ -1,5 +1,7 @@
 import re
 
+from sanic.request import File
+
 from insanic.conf import settings
 
 HIDDEN_KEY_WORDS = [
@@ -9,12 +11,15 @@ HIDDEN_KEY_WORDS = [
     'SECRET',
     'PASS',
     'PROFANITIES_LIST',
-    'SIGNATURE'
+    'SIGNATURE',
+    'SESSION',
+    "EMAIL",
+    "PHONE"
 ]
 
 HIDDEN_SETTINGS = re.compile("|".join(HIDDEN_KEY_WORDS + [v.lower() for v in HIDDEN_KEY_WORDS]))
 
-CLEANSED_SUBSTITUTE = '********************'
+CLEANSED_SUBSTITUTE = '*********'
 
 def tracing_name(name=None):
     """
@@ -53,6 +58,15 @@ def get_safe_dict(target):
     for k in target:
         return_value[k] = cleanse_value(k, target.get(k))
     return return_value
+
+
+def abbreviate_for_xray(payload):
+    for k in payload.keys():
+        v = payload.get(k)
+        if isinstance(v, File):
+            v = {"type": v.type, "size": len(v.body)}
+        payload[k] = v
+    return payload
 
 
 def get_safe_settings():